mainwindow.ui: example .ui for a pyside-uic bug
Running the command below with pyside-uic generates a bug in pyside-uic

$ pyside-uic mainwindow.ui -o ui_mainwindow.py
An unexpected error occurred.
Check that you are using the latest version of PySide and report the error to
http://bugs.openbossa.org, including the ui file used to trigger the error.

$ pyside-uic --version
PySide User Interface Compiler version 0.2.13, running on PySide 1.1.1.

The difference from the parent commit is that a new save item was added, and
the theme document-save icon was assigned to this new item instead of the
previous one.